clean up aa endpts and limit ffprobe fields
