--- mediorum/server/serve_audio_analysis.go
@@ -1,25 +1,0 @@
-package server
-
-import (
-	"github.com/labstack/echo/v4"
-)
-
-// returns analysis results if analysis previously succeeded.
-func (ss *MediorumServer) analyzeUpload(c echo.Context) error {
-	var upload *Upload
-	err := ss.crud.DB.First(&upload, "id = ?", c.Param("id")).Error
-	if err != nil {
-		return echo.NewHTTPError(404, err.Error())
-	}
-	return c.JSON(200, upload)
-}
-
-func (ss *MediorumServer) serveLegacyBlobAnalysis(c echo.Context) error {
-	cid := c.Param("cid")
-	var analysis *QmAudioAnalysis
-	err := ss.crud.DB.First(&analysis, "cid = ?", cid).Error
-	if err != nil {
-		return echo.NewHTTPError(404, err.Error())
-	}
-	return c.JSON(200, analysis)
-}
