+++ mediorum/server/server.go
@@ -299,7 +299,6 @@ func New(config MediorumConfig) (*MediorumServer, error) {
 	routes.GET("/uploads", ss.serveUploadList)
 	routes.GET("/uploads/:id", ss.serveUploadDetail, ss.requireHealthy)
 	routes.POST("/uploads/:id", ss.updateUpload, ss.requireHealthy, ss.requireUserSignature)
-	routes.POST("/uploads/:id/analyze", ss.analyzeUpload, ss.requireHealthy, ss.requireRegisteredSignature)
 	routes.POST("/uploads", ss.postUpload, ss.requireHealthy)
 	// workaround because reverse proxy catches the browser's preflight OPTIONS request instead of letting our CORS middleware handle it
 	routes.OPTIONS("/uploads", func(c echo.Context) error {
@@ -308,7 +307,6 @@ func New(config MediorumConfig) (*MediorumServer, error) {
 
 	// legacy blob audio analysis
 	routes.GET("/tracks/legacy/:cid/analysis", ss.serveLegacyBlobAnalysis, ss.requireHealthy)
-	routes.POST("/tracks/legacy/:cid/analyze", ss.serveLegacyBlobAnalysis, ss.requireHealthy)
 
 	// serve blob (audio)
 	routes.HEAD("/ipfs/:cid", ss.serveBlob, ss.requireHealthy, ss.ensureNotDelisted)
